requirements.txt
# -*- coding: utf-8 -*-
dataclasses-json==0.5.2
google-api-core==1.13.0
googleapis-common-protos==1.5.10
grpcio==1.31.0
grpcio-reflection==1.31.0
grpcio-tools==1.31.0
mypy-protobuf==1.20
polling==0.3.1
pytest==6.0.1
regex==2020.7.14
protobuf==3.13.0
